+++ packages/terra-modal-manager/src/terra-dev-site/doc/example/ModalManagerExample.jsx
@@ -93,7 +93,7 @@ class DisclosureComponent extends React.Component {
                 });
             }}
           >
-          Dismiss
+            Dismiss
           </button>
           <button
             type="button"
@@ -108,7 +108,7 @@ class DisclosureComponent extends React.Component {
               });
             }}
           >
-          Disclose Again
+            Disclose Again
           </button>
           <br />
           <br />