+++ packages/terra-slide-panel-manager/src/terra-dev-site/doc/example/SlidePanelManagerExample.jsx
@@ -94,7 +94,7 @@ class DisclosureComponent extends React.Component {
                 });
             }}
           >
-          Dismiss
+            Dismiss
           </button>
           <button
             type="button"
@@ -109,7 +109,7 @@ class DisclosureComponent extends React.Component {
               });
             }}
           >
-          Disclose Again
+            Disclose Again
           </button>
           <br />
           <br />