+++ packages/terra-time-input/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+
+4.30.0 - (July 14, 2020)
+------------------
 ### Changed
 * Minor file changes to comply with eslint rules
 * Remove unused disable eslint directives